third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/reduce_product.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] reduceProduct float16 0D constant tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test reduceProduct float16 actual 0 should be close enough to expected -68.75 by ULP distance: expected a number less than or equal to 1 but got 21580
[FAIL] reduceProduct float16 0D constant tensor empty axes
  assert_less_than_equal: assert_array_approx_equals_ulp: test reduceProduct float16 actual 0 should be close enough to expected -68.75 by ULP distance: expected a number less than or equal to 1 but got 21580
[FAIL] reduceProduct float16 1D constant tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test reduceProduct float16 actual -2 should be close enough to expected -0.1248779296875 by ULP distance: expected a number less than or equal to 24 but got 4098
[FAIL] reduceProduct float16 1D tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test reduceProduct float16 actual 0 should be close enough to expected -0.1248779296875 by ULP distance: expected a number less than or equal to 24 but got 45054
[FAIL] reduceProduct float16 2D tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test reduceProduct float16 actual 0 should be close enough to expected -0.1248779296875 by ULP distance: expected a number less than or equal to 24 but got 45054
[FAIL] reduceProduct float16 3D tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test reduceProduct float16 actual 0 should be close enough to expected -0.1248779296875 by ULP distance: expected a number less than or equal to 24 but got 45054
[FAIL] reduceProduct float16 4D tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test reduceProduct float16 actual 0 should be close enough to expected -0.1248779296875 by ULP distance: expected a number less than or equal to 24 but got 45054
[FAIL] reduceProduct float16 5D tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test reduceProduct float16 actual -2 should be close enough to expected -0.1248779296875 by ULP distance: expected a number less than or equal to 24 but got 4098
[FAIL] reduceProduct float16 3D tensor options.axes
  assert_less_than_equal: assert_array_approx_equals_ulp: test reduceProduct float16 actual 0 should be close enough to expected -0.3759765625 by ULP distance: expected a number less than or equal to 4 but got 46596
[FAIL] reduceProduct float16 4D tensor options.axes
  assert_less_than_equal: assert_array_approx_equals_ulp: test reduceProduct float16 actual 0 should be close enough to expected 1.2744140625 by ULP distance: expected a number less than or equal to 4 but got 15641
[FAIL] reduceProduct float16 3D tensor options.keepDimensions=false
  assert_less_than_equal: assert_array_approx_equals_ulp: test reduceProduct float16 actual 0 should be close enough to expected -0.1248779296875 by ULP distance: expected a number less than or equal to 24 but got 45054
[FAIL] reduceProduct float16 3D tensor options.keepDimensions=true
  assert_less_than_equal: assert_array_approx_equals_ulp: test reduceProduct float16 actual 0 should be close enough to expected -0.1248779296875 by ULP distance: expected a number less than or equal to 24 but got 45054
[FAIL] reduceProduct float16 4D tensor options.keepDimensions=false
  assert_less_than_equal: assert_array_approx_equals_ulp: test reduceProduct float16 actual 0 should be close enough to expected -0.1248779296875 by ULP distance: expected a number less than or equal to 24 but got 45054
[FAIL] reduceProduct float16 4D tensor options.keepDimensions=true
  assert_less_than_equal: assert_array_approx_equals_ulp: test reduceProduct float16 actual 0 should be close enough to expected -0.1248779296875 by ULP distance: expected a number less than or equal to 24 but got 45054
[FAIL] reduceProduct float16 4D tensor options.axes with options.keepDimensions=false
  assert_less_than_equal: assert_array_approx_equals_ulp: test reduceProduct float16 actual 0 should be close enough to expected 1.666015625 by ULP distance: expected a number less than or equal to 6 but got 16042
[FAIL] reduceProduct float16 4D tensor options.axes with options.keepDimensions=true
  assert_less_than_equal: assert_array_approx_equals_ulp: test reduceProduct float16 actual 0 should be close enough to expected 1.666015625 by ULP distance: expected a number less than or equal to 6 but got 16042
Harness: the test ran to completion.
